+++ src/api/types/application-types/index.ts
@@ -111,7 +111,15 @@ export interface ApplicationDeclaration extends Relationship {
   hasAntiBriberyCodeOfConduct?: boolean | null;
   willExportWithAntiBriberyCodeOfConduct?: boolean;
   agreeToConfirmationAndAcknowledgements?: boolean;
-  agreeHowDataWillBeUsed?: boolean;
+}
+
+export interface ApplicationDeclarationVersions {
+  ANTI_BRIBERY: string;
+  ANTI_BRIBERY_CODE_OF_CONDUCT: string;
+  ANTI_BRIBERY_EXPORTING_WITH_CODE_OF_CONDUCT: string;
+  CONFIDENTIALITY: string;
+  CONFIRMATION_AND_ACKNOWLEDGEMENTS: string;
+  HOW_YOUR_DATA_WILL_BE_USED?: string;
 }
 
 export interface ApplicationLossPayeeFinancialInternationalVector extends Relationship {