+++ src/ui/server/constants/declarations/versions/index.ts
@@ -1,27 +1,27 @@
-import { ApplicationDeclarationVersion } from '../../../../types';
+import { ApplicationDeclarationVersions } from '../../../../types';
 
 /**
- * VERSIONS
+ * DECLARATION_VERSIONS
  * All possible declaration versions.
  * This should be manually updated each time declarations are updated
- * @returns {Array<ApplicationDeclarationVersion>} All declaration versions
+ * @returns {Array<ApplicationDeclarationVersions>} All declaration versions
  */
-const VERSIONS = [
+const DECLARATION_VERSIONS = [
   {
-    CONFIDENTIALITY: '1',
     ANTI_BRIBERY: '1',
     ANTI_BRIBERY_CODE_OF_CONDUCT: '1',
     ANTI_BRIBERY_EXPORTING_WITH_CODE_OF_CONDUCT: '1',
+    CONFIDENTIALITY: '1',
     CONFIRMATION_AND_ACKNOWLEDGEMENTS: '1',
     HOW_YOUR_DATA_WILL_BE_USED: '1',
   },
   {
-    CONFIDENTIALITY: '1',
     ANTI_BRIBERY: '1',
     ANTI_BRIBERY_CODE_OF_CONDUCT: '1',
     ANTI_BRIBERY_EXPORTING_WITH_CODE_OF_CONDUCT: '1',
+    CONFIDENTIALITY: '1',
     CONFIRMATION_AND_ACKNOWLEDGEMENTS: '1',
   },
-] as Array<ApplicationDeclarationVersion>;
+] as Array<ApplicationDeclarationVersions>;
 
-export default VERSIONS;
+export default DECLARATION_VERSIONS;