+++ src/ui/server/graphql/queries/application-by-reference-number.ts
@@ -227,7 +227,6 @@ const applicationByReferenceNumberQuery = gql`
           hasAntiBriberyCodeOfConduct
           willExportWithAntiBriberyCodeOfConduct
           agreeToConfirmationAndAcknowledgements
-          agreeHowDataWillBeUsed
         }
       }
     }