+++ src/ui/types/application.ts
@@ -189,9 +189,18 @@ interface ApplicationDeclaration {
   id: string;
   agreeToConfidentiality?: boolean;
   agreeToAntiBribery?: boolean;
-  hasAntiBriberyCodeOfConduct?: boolean | null;
+  hasAntiBriberyCodeOfConduct?: boolean;
   willExportWithAntiBriberyCodeOfConduct?: boolean;
   agreeToConfirmationAndAcknowledgements?: boolean;
+}
+
+interface ApplicationDeclarationVersions {
+  ANTI_BRIBERY: string;
+  ANTI_BRIBERY_CODE_OF_CONDUCT: string;
+  ANTI_BRIBERY_EXPORTING_WITH_CODE_OF_CONDUCT: string;
+  CONFIDENTIALITY: string;
+  CONFIRMATION_AND_ACKNOWLEDGEMENTS: string;
+  HOW_YOUR_DATA_WILL_BE_USED?: string;
 }
 
 interface ApplicationPolicyContact {
@@ -283,15 +292,6 @@ interface ApplicationVersion {
   DEFAULT_NEED_PRE_CREDIT_PERIOD_COVER: boolean;
 }
 
-interface ApplicationDeclarationVersion {
-  CONFIDENTIALITY: string;
-  ANTI_BRIBERY: string;
-  ANTI_BRIBERY_CODE_OF_CONDUCT: string;
-  ANTI_BRIBERY_EXPORTING_WITH_CODE_OF_CONDUCT: string;
-  CONFIRMATION_AND_ACKNOWLEDGEMENTS: string;
-  HOW_YOUR_DATA_WILL_BE_USED?: string;
-}
-
 export {
   Application,
   ApplicationBusiness,
@@ -303,6 +303,8 @@ export {
   ApplicationBuyerApiInput,
   ApplicationCompany,
   ApplicationCompanyDifferentTradingAddress,
+  ApplicationDeclaration,
+  ApplicationDeclarationVersions,
   ApplicationExportContract,
   ApplicationExportContractAgent,
   ApplicationExportContractAgentService,
@@ -317,8 +319,6 @@ export {
   ApplicationPolicy,
   ApplicationPolicyContact,
   ApplicationSectionReview,
-  ApplicationDeclaration,
-  ApplicationDeclarationVersion,
-  ApplicationVersion,
   ApplicationOwner,
+  ApplicationVersion,
 };